requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
rank-bm25==0.2.2
pydantic==2.5.0
python-multipart==0.0.6
faiss-cpu==1.7.4
rank-bm25==0.2.2
pydantic==2.5.0
python-multipart==0.0.6
